+++ googlesource/source_intergration_test.go
@@ -26,7 +26,6 @@ import (
 	sdk "github.com/conduitio/conduit-connector-sdk"
 	googlebigquery "github.com/neha-Gupta1/conduit-connector-bigquery"
 	"google.golang.org/api/option"
-	"gopkg.in/tomb.v2"
 )
 
 var (
@@ -146,18 +145,6 @@ func cleanupDataSet() (err error) {
 	return err
 }
 
-func TestConfigureSource_FailsWhenConfigEmpty(t *testing.T) {
-	con := Source{}
-	err := con.Configure(context.Background(), make(map[string]string))
-	if err == nil {
-		t.Errorf("expected no error, got %v", err)
-	}
-
-	if strings.HasPrefix(err.Error(), "config is invalid:") {
-		t.Errorf("expected error to be about missing config, got %v", err)
-	}
-}
-
 func TestSuccessfulGet(t *testing.T) {
 	// cleanupDataSet()
 	err := dataSetup()
@@ -260,129 +247,5 @@ func TestSuccessfulGetWholeDataset(t *testing.T) {
 		fmt.Println("Record found:", value)
 		value = string(record.Payload.Bytes())
 		fmt.Println(":", value)
-	}
-	// fmt.Println("Calling teardown for start...")
-	// err = src.Teardown(ctx)
-	// fmt.Println("Calling teardown...")
-	// if err != nil {
-	// 	t.Errorf("expected no error, got %v", err)
-
-	// }
-}
-
-func TestSuccessfulTearDown(t *testing.T) {
-
-	err := dataSetup()
-	if err != nil {
-		fmt.Println("Could not create values. Err: ", err)
-		return
-	}
-	defer cleanupDataSet()
-
-	src := Source{}
-	cfg := map[string]string{}
-
-	cfg[googlebigquery.ConfigServiceAccount] = serviceAccount
-	cfg[googlebigquery.ConfigProjectID] = projectID
-	cfg[googlebigquery.ConfigDatasetID] = datasetID
-	cfg[googlebigquery.ConfigTableID] = tableID
-
-	ctx := context.Background()
-	err = src.Configure(ctx, cfg)
-	if err != nil {
-		fmt.Println(err)
-	}
-
-	pos := sdk.Position{}
-	err = src.Open(ctx, pos)
-	if err != nil {
-		fmt.Println("errror: ", err)
-	}
-
-	err = src.Teardown(ctx)
-	if err != nil {
-		t.Errorf("expected no error, got %v", err)
-
-	}
-}
-
-func TestInvalidCreds(t *testing.T) {
-
-	src := Source{}
-	cfg := map[string]string{}
-	cfg[googlebigquery.ConfigServiceAccount] = "invalid"
-	cfg[googlebigquery.ConfigProjectID] = projectID
-	cfg[googlebigquery.ConfigDatasetID] = datasetID
-	cfg[googlebigquery.ConfigTableID] = tableID
-	cfg[googlebigquery.ConfigLocation] = "test"
-
-	googlebigquery.PollingTime = time.Second * 1
-	ctx := context.Background()
-	err := src.Configure(ctx, cfg)
-	if err != nil {
-		fmt.Println(err)
-	}
-
-	pos := sdk.Position{}
-	err = src.Open(ctx, pos)
-	if err != nil {
-		t.Errorf("expected error, got nil")
-	}
-
-	for {
-
-		record, err := src.Read(ctx)
-		if err != nil && err == sdk.ErrBackoffRetry {
-			fmt.Println("err: ", err)
-			break
-		}
-		if err != nil {
-			t.Errorf("some other error found: %v", err)
-		}
-		value := string(record.Position)
-		fmt.Println("Record found:", value)
-		value = string(record.Payload.Bytes())
-		fmt.Println(":", value)
-	}
-
-}
-
-func TestNewSource(t *testing.T) {
-	NewSource()
-}
-
-func TestAck(t *testing.T) {
-	s := NewSource()
-	s.Ack(context.TODO(), sdk.Position{})
-}
-
-func TestTableFetchInvalidCred(t *testing.T) {
-	s := Source{}
-	_, err := s.listTables("", "")
-	if err == nil {
-		t.Errorf("expected error, got nil")
-	}
-}
-
-// func TestValueFromTypeMap(t *testing.T) {
-
-// 	type testStruct struct {
-// 		name  string
-// 		value *int
-// 	}
-// 	value := 1
-// 	valueFromTypeMap(testStruct{name: "Test", value: &value})
-// }
-
-func TestNextContextDone(t *testing.T) {
-
-	s := Source{}
-	ctx := context.Background()
-	ctx, cancel := context.WithCancel(ctx)
-	cancel()
-	s.tomb = &tomb.Tomb{}
-	_, err := s.Next(ctx)
-	if err == nil {
-		t.Errorf("expected error, got nil")
 	}
 }
